+++ packages/sample-app/src/@types/dynamic-plugin-sdk.d.ts
@@ -1,0 +1,10 @@
+import '@openshift/dynamic-plugin-sdk';
+
+declare module '@openshift/dynamic-plugin-sdk' {
+  interface PluginCustomProperties {
+    sampleApp: {
+      /** What kind of lunch the plugin author wants to eat */
+      lunch: string;
+    };
+  }
+}